requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]